# docs/swagger.yaml
basePath: /
definitions:
...
  bson.M:
    additionalProperties: {}
    type: object
...
  handlers.WorkflowListResponse:
    properties:
      has_more:
        type: boolean
      next_cursor:
        type: string
      workflows:
        items:
          $ref: '#/definitions/mongostore.Workflow'
        type: array
    type: object
...
  mongostore.Workflow:
    properties:
      created_at:
        type: string
      description:
        type: string
      document_id:
        type: string
      edges:
        items:
          $ref: '#/definitions/bson.M'
        type: array
      id:
        type: string
      name:
        type: string
      nodes:
        description: |-
          React Flow's own toObject() shape, stored as-is and re-rendered with
          whatever the current node/edge components are. Deliberately untyped
          (bson.M, not a Go struct): the node/edge schema is still being
          designed on the frontend and is expected to keep changing without a
          backend release in lockstep — see the no-migration-logic note on
          reading old versions.
        items:
          $ref: '#/definitions/bson.M'
        type: array
      tags:
        items:
          type: string
        type: array
      version:
        type: integer
      viewport:
        $ref: '#/definitions/bson.M'
    type: object
...
info:
  contact: {}
  description: Event-driven Go API backed by MongoDB and Redis. Every request carries
    a correlation ID (X-Correlation-ID), generated by the server if the caller doesn't
    supply one.
  title: Metarr API
  version: "1.0"
paths:
...
  /api/workflows:
    get:
      description: Returns the latest version of every saved workflow, newest first,
        paginated by an opaque cursor.
      parameters:
      - description: Maximum records to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Opaque cursor from a previous page's next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkflowListResponse'
        "400":
          description: invalid limit or cursor
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: List workflows
      tags:
      - Workflows
    post:
      consumes:
      - application/json
      description: Creates a new workflow, or appends a new version to an existing
        one if document_id is set. Every save is a new version — nothing is overwritten
        in place.
      parameters:
      - description: Workflow to save
        in: body
        name: workflow
        required: true
        schema:
          $ref: '#/definitions/mongostore.Workflow'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/mongostore.Workflow'
        "400":
          description: malformed body, or missing name/description/tags
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Save a workflow
      tags:
      - Workflows
  /api/workflows/{id}:
    get:
      description: Returns the latest version of one workflow.
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mongostore.Workflow'
        "400":
          description: malformed id
          schema:
            type: string
        "404":
          description: no workflow with that id
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch a workflow
      tags:
      - Workflows
  /api/workflows/{id}/versions:
    get:
      description: Returns every saved version of one workflow, newest first, for
        the version-history strip.
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mongostore.Workflow'
            type: array
        "400":
          description: malformed id
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: List a workflow's versions
      tags:
      - Workflows
  /api/workflows/{id}/versions/{version}:
    get:
      description: Returns one specific past version of a workflow, exactly as it
        was saved.
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mongostore.Workflow'
        "400":
          description: malformed id or version
          schema:
            type: string
        "404":
          description: no such workflow or version
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch one workflow version
      tags:
      - Workflows
securityDefinitions:
  ApiKeyHeaderAuth:
    description: API key via the X-Api-Key header.
    in: header
    name: X-Api-Key
    type: apiKey
  ApiKeyQueryAuth:
    description: API key via the apikey query parameter. Either this or ApiKeyHeaderAuth
      satisfies auth on protected routes.
    in: query
    name: apikey
    type: apiKey
swagger: "2.0"